ResumeAnalyzer/pyproject.toml
[project]
name = "resumeanalyzer"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "chromadb>=1.5.1",
    "faiss-cpu>=1.13.2",
    "fastapi>=0.135.1",
    "langchain>=1.2.10",
    "langchain-community>=0.4.1",
    "langchain-core>=1.2.15",
    "langchain-groq>=1.1.2",
    "pdfplumber>=0.11.9",
    "pydantic>=2.12.5",
    "pymupdf>=1.27.1",
    "pypdf>=6.7.3",
    "python-dotenv>=1.2.1",
    "python-multipart>=0.0.22",
    "sentence-transformers>=5.2.3",
    "uvicorn>=0.41.0",
]
